msvc builds failed due to some narrowing conversion?

+++ apps/bench/OSPRayFixture.cpp
@@ -32,8 +32,8 @@ std::vector<string> OSPRayFixture::benchmarkModelFiles;
 int OSPRayFixture::width  = 1024;
 int OSPRayFixture::height = 1024;
 
-int OSPRayFixture::numBenchFrames = 100;
-int OSPRayFixture::numWarmupFrames = 10;
+size_t OSPRayFixture::numBenchFrames = 100;
+size_t OSPRayFixture::numWarmupFrames = 10;
 bool OSPRayFixture::logFrameTimes = false;
 
 vec3f OSPRayFixture::bg_color = {1.f, 1.f, 1.f};
